Event listener to all calculator buttons
When a button number is "clicked", the display innerText will change

When the user clicks on an operation button, the first number will be stored
on the displayOperation variable. The math operation will be stored as well

When the user clicks the result button, the second number will be stored
on the displayOperation variable

This is the first step in making the calculator work. The calculator is still not
calling the math function.